Move toolregistry package to pkg/plugin
