+++ pkg/plugin/toolregistry/toolregistry.go
