+++ pkg/plugin/toolregistry/toolregistrytest/toolregistrytest.go
